+++ tests/update_banner.rs
@@ -44,6 +44,7 @@ impl ContentProvider for Content {
             content: Text::raw(""),
             notices: Vec::new(),
             source: None,
+            image: None,
         }
     }
 }